Edit prods / Presupuestos

+++ src/admin/componentes/Productos/addProds.js
@@ -6,16 +6,16 @@ import { ref, uploadBytesResumable, getDownloadURL } from "firebase/storage";
 import Swal from 'sweetalert2';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-
+import { FadeLoader } from 'react-spinners';
 
 function AddProds() {
-    const { register, handleSubmit,  watch  , formState: { errors } } = useForm()
+    const { register, handleSubmit, watch, formState: { errors } } = useForm()
     const navigate = useNavigate()
-
+    const [isLoading, setIsLoading] = useState(false)
     const [products, setProducts] = useState([]);
     const [imageFile, setImageFile] = useState(null);
 
-    
+
     const handleNavigateToEdit = (productId) => {
         navigate(`/admin/editProds/${productId}`);
 
@@ -26,8 +26,10 @@ function AddProds() {
         precio: '',
         descripcion: '',
         imagen: '',
-        //... otros campos que quieras
-    });
+        stock: '',
+        talle: '',
+        categoria: '',
+        });
 
     const fetchProducts = async () => {
         const productsCollection = query(collection(baseDeDatos, "productos"));
@@ -35,7 +37,7 @@ function AddProds() {
         const productList = productSnapshot.docs.map(doc => ({ ...doc.data(), id: doc.id }));
         setProducts(productList);
     };
-    
+
     useEffect(() => {
         fetchProducts();
     }, []);
@@ -47,15 +49,17 @@ function AddProds() {
     };
 
     const onSubmit = async (e) => {
-  
-        const fieldsFilled = (                            
+
+        setIsLoading(true)
+
+        const fieldsFilled = (
             watch('nombreProd') &&
             watch('precioProd') &&
             watch('descrProd') &&
             watch('imagenProd') &&
             watch('categoria') &&
             watch('stock') &&
-            watch('size')         
+            watch('talle')
         );
         if (!fieldsFilled) {
             Swal.fire({
@@ -65,21 +69,21 @@ function AddProds() {
             });
             return;  // Salir de la función
         }
-      
+        
         try {
             // Subir la imagen a Firebase Storage
             if (imageFile) {
                 const storageRef = ref(storage, 'productos/' + imageFile.name);
                 const uploadTask = uploadBytesResumable(storageRef, imageFile);
-    
-                uploadTask.on('state_changed', 
+
+                uploadTask.on('state_changed',
                     (snapshot) => {
                         // Progreso del upload
-                    }, 
+                    },
                     (error) => {
                         // Error en el upload
                         console.error(error);
-                    }, 
+                    },
                     async () => {
                         // Upload completado exitosamente, obtener la URL de descarga
                         const downloadURL = await getDownloadURL(uploadTask.snapshot.ref);
@@ -88,65 +92,76 @@ function AddProds() {
                             ...productData,
                             imagen: downloadURL
                         };
-                        
+
                         // Ahora puedes agregar el producto a Firestore
                         const docRef = await addDoc(collection(baseDeDatos, "productos"), productDataWithImage);
                         console.log("Producto agregado con foto y con ID: ", docRef.id);
+
+                        Swal.fire({
+                            icon: 'success',
+                            title: 'Producto Agregado',
+                            text: `Has agregado correctamente un nuevo Producto`,
+                        });
+                        fetchProducts()
                         
                         // Luego de agregar el producto puedes limpiar el estado
                         setProductData({
                             nombre: '',
                             precio: '',
                             descripcion: '',
+                            stock: '',
+                            talle: '',
+                            categoria: '',
                             imagen: '',
                         });
                     }
                 );
-                Swal.fire({
-                    icon: 'success',
-                    title: 'Producto Agregado',
-                    text: `Has agregado correctamente un nuevo Producto`,
-                  });
-                } 
-            fetchProducts()
+               
+            }
+            setIsLoading(false)
         } catch (e) {
+            setIsLoading(false)
             console.error("Error al añadir el producto: ", e);
         }
     };
-    
+
 
     const deleteProduct = async (productId) => {
         try {
-            await deleteDoc(doc(baseDeDatos, "productos", productId));
-            console.log("Producto eliminado");
-            Swal.fire({
-                icon: 'success',
-                title: 'Producto Eliminado',
-                text: `Has eliminado un producto`,
-              });
-            fetchProducts()
+
+            const result = await Swal.fire({
+                title: '¿Estás seguro?',
+                text: "No podrás revertir esta acción.",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Sí, eliminar',
+                cancelButtonText: 'Cancelar'
+            });
+            // Si el usuario confirmó la eliminación
+            if (result.isConfirmed) {
+                // Aquí colocas tu código para eliminar el producto
+               
+                    await deleteDoc(doc(baseDeDatos, "productos", productId));
+                    console.log("Producto eliminado");
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Producto Eliminado',
+                        text: `Has eliminado un producto`,
+                    });
+                    fetchProducts()
+                }
+
         } catch (e) {
             console.error("Error al eliminar el producto: ", e);
+            Swal.fire(
+                'Error',
+                `Hubo un problema eliminando el producto. Error:${e}`,
+                'error'
+            );
         }
     }
-
-    const updateProduct = async (productId, updatedData) => {
-        try {
-            const productRef = doc(baseDeDatos, "productos", productId);
-            await updateDoc(productRef, updatedData);
-            console.log("Producto actualizado");
-            Swal.fire({
-                icon: 'success',
-                title: 'Producto Editado',
-                text: `Has editado correctamente un Producto `,
-              });
-            fetchProducts()
-        } catch (e) {
-            console.error("Error al actualizar el producto: ", e);
-        }
-    }
-
-
 
     return (
         < div className='div-add-edit-prods'>
@@ -157,83 +172,91 @@ function AddProds() {
                 <form className='form-addProd' onSubmit={handleSubmit(onSubmit)}>
                     <label> Nombre del producto </label>
                     <input
-                    {...register("nombreProd", { required: true })}
+                        {...register("nombreProd", { required: true })}
                         value={productData.nombre}
                         name="nombreProd"
 
                         onChange={e => setProductData({ ...productData, nombre: e.target.value })}
                         placeholder="Nombre del producto"
                     />
-                  {errors.nombreUser && <p className='message-error' > El nombre del producto es requerido</p>}
+                    {errors.nombreProd && <p className='message-error' > El nombre del producto es requerido</p>}
 
-                  <label> Categoria </label>
+                    <label> Categoria </label>
                     <input
-                    {...register("categoria", { required: true })}
+                        {...register("categoria", { required: true })}
                         value={productData.categoria}
                         name="categoria"
 
                         onChange={e => setProductData({ ...productData, categoria: e.target.value })}
                         placeholder="Categoria del producto"
                     />
-                  {errors.nombreUser && <p className='message-error' > la categoria del producto es requerida</p>}
+                    {errors.categoria && <p className='message-error' > la categoria del producto es requerida</p>}
 
 
 
                     <label> Precio del producto </label>
                     <input
-                    {...register("precioProd", { required: true })}
+                        {...register("precioProd", { required: true })}
                         value={productData.precio}
                         name="precioProd"
 
                         onChange={e => setProductData({ ...productData, precio: e.target.value })}
                         placeholder="Precio del producto"
                     />
-                    {errors.nombreUser && <p className='message-error' > El precio del producto es requerido</p>}
+                    {errors.precioProd && <p className='message-error' > El precio del producto es requerido</p>}
 
                     <label> Descripción del producto </label>
                     <textarea
-                    {...register("descrProd", { required: true })}
+                        {...register("descrProd", { required: true })}
                         value={productData.descripcion}
                         name="descrProd"
                         onChange={e => setProductData({ ...productData, descripcion: e.target.value })}
                         placeholder="Descripción del producto"
                     />
-                    {errors.nombreUser && <p className='message-error' > Agregue una descripcion del producto</p>}
+                    {errors.descrProd && <p className='message-error' > Agregue una descripcion del producto</p>}
 
 
                     <label> Stock </label>
                     <input
-                    {...register("stock", { required: true })}
+                        {...register("stock", { required: true })}
                         value={productData.stock}
                         name="stock"
                         type='number'
                         onChange={e => setProductData({ ...productData, stock: e.target.value })}
                         placeholder="Stock del producto"
                     />
-                    {errors.nombreUser && <p className='message-error' > Agregue un Stock al producto</p>}
+                    {errors.stock && <p className='message-error' > Agregue un Stock al producto</p>}
 
 
                     <label> Talle del producto </label>
                     <input
-                    {...register("size", { required: true })}
-                        value={productData.size}
-                        name="size"
-                        onChange={e => setProductData({ ...productData, size: e.target.value })}
+                        {...register("talle", { required: true })}
+                        value={productData.talle}
+                        name="talle"
+                        onChange={e => setProductData({ ...productData, talle: e.target.value })}
                         placeholder="Talle del producto"
                     />
-                    {errors.nombreUser && <p className='message-error' > Agregue un talle al producto</p>}
+                    {errors.talle && <p className='message-error' > Agregue un talle al producto</p>}
 
 
                     <label> Imagen </label>
                     <input
-                    {...register("imagenProd", { required: true })}
+                        {...register("imagenProd", { required: true })}
                         type="file"
                         name="imagenProd"
                         onChange={e => handleFileChange(e)}
                     />
-                    {errors.nombreUser && <p className='message-error' > Debe agregar una Imagen</p>}
+                    {errors.imagenProd && <p className='message-error' > Debe agregar una Imagen</p>}
 
-                    <button className='add-prod-btn' type="submit">Agregar Producto</button>
+
+                    {
+                        isLoading ? (
+                            <div className="spinner">
+                                Agregando Nuevo Producto, aguarde....
+                                <FadeLoader color="black" />
+                            </div>
+                        ) : <button className='add-prod-btn' type="submit">Agregar Producto</button>
+                    }
                 </form>
             </div>
 
@@ -259,7 +282,7 @@ function AddProds() {
                                 <td className="td-tbody"><img src={product.imagen} alt={product.nombre} width="50" /></td>
                                 <td className="td-tbody">
                                     <div className='btns-table'>
-                                    <button className="btn-table-edit"   onClick={() => handleNavigateToEdit(product.id)}>Editar</button>
+                                        <button className="btn-table-edit" onClick={() => handleNavigateToEdit(product.id)}>Editar</button>
 
                                         {/* <button className="btn-table-edit" onClick={() => updateProduct(product.id, product)}>Editar</button> */}
                                         <button className="btn-table-delete" onClick={() => deleteProduct(product.id)}>Eliminar</button>
